# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for CompStats - stat tracking for online melee game tournaments.",
        "title": "CompStats API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/admin/players": {
            "get": {
                "description": "get all players",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Players",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/database.Player"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "create a player",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create player",
                "parameters": [
                    {
                        "description": "create json",
                        "name": "dto",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/player.createDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/database.Player"
                        }
                    }
                }
            }
        },
        "/admin/players/{id}": {
            "get": {
                "description": "get player by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Get a player",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Player ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/database.Player"
                        }
                    }
                }
            },
            "delete": {
                "description": "delete player by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Delete a player",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Player ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "patch": {
                "description": "update player by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Update a player",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Player ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "update json",
                        "name": "dto",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/player.updateDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/database.Player"
                        }
                    }
                }
            }
        },
        "/admin/tournaments": {
            "get": {
                "description": "get all tournaments",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Tournaments",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/database.Tournament"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "create a tournament",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create tournament",
                "parameters": [
                    {
                        "description": "create json",
                        "name": "dto",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/tournament.createUpdateDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/database.Tournament"
                        }
                    }
                }
            }
        },
        "/admin/tournaments/{id}": {
            "get": {
                "description": "get tournament by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Get a tournament",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Tournament ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/database.Tournament"
                        }
                    }
                }
            },
            "delete": {
                "description": "delete tournament by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Delete a tournament",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Tournament ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "patch": {
                "description": "update tournament by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Update a tournament",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Tournament ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "update json",
                        "name": "dto",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/tournament.createUpdateDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/database.Tournament"
                        }
                    }
                }
            }
        },
...
        "/stats": {
            "post": {
                "description": "save half, players and stats from game server",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Update stats",
                "parameters": [
                    {
                        "description": "update stats json",
                        "name": "dto",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/player.updateStatsDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/player.updateStatsResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "database.Player": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "virtualId": {
                    "type": "string"
                }
            }
        },
        "database.Tournament": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "player.createDTO": {
            "type": "object",
            "required": [
                "name",
                "virtual_id"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "virtual_id": {
                    "type": "string"
                }
            }
        },
        "player.playerStats": {
            "type": "object",
            "required": [
                "assists",
                "deaths",
                "kills",
                "score",
                "win"
            ],
            "properties": {
                "assists": {
                    "type": "integer"
                },
                "deaths": {
                    "type": "integer"
                },
                "kills": {
                    "type": "integer"
                },
                "score": {
                    "type": "integer"
                },
                "win": {
                    "type": "boolean"
                }
            }
        },
        "player.playerToBeUpdated": {
            "type": "object",
            "required": [
                "name",
                "stats",
                "virtual_id"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "stats": {
                    "$ref": "#/definitions/player.playerStats"
                },
                "virtual_id": {
                    "type": "string"
                }
            }
        },
        "player.updateDTO": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "virtual_id": {
                    "type": "string"
                }
            }
        },
        "player.updateStatsDTO": {
            "type": "object",
            "required": [
                "gameType",
                "mapName",
                "players",
                "team1",
                "team2",
                "tournamentName"
            ],
            "properties": {
                "gameType": {
                    "type": "string"
                },
                "mapName": {
                    "type": "string"
                },
                "players": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/player.playerToBeUpdated"
                    }
                },
                "team1": {
                    "type": "string"
                },
                "team2": {
                    "type": "string"
                },
                "tournamentName": {
                    "type": "string"
                }
            }
        },
        "player.updateStatsResponse": {
            "type": "object",
            "required": [
                "message"
            ],
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "tournament.createUpdateDTO": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        }
    }
}